Refactor wine data processing and update project settings
- Consolidated string values in `OperationsEnums.cs` for improved readability and clarified comments regarding case sensitivity of `Enum.Parse` and `Enum.TryParse`.
- Updated method calls in `Program.cs` to change execution flow.
- Modified `WineContext.cs` to use a dynamic connection string and conditional data seeding.
- Changed target framework to .NET 9.0 in `EnumHasConversionSample.csproj` and added new package references.
- Introduced a new initialization method in `Program.cs` for console setup and dependency injection.
- Added `Utilities.cs` for database existence checks using Dapper.
- Created `appsettings.json` for configuration management.
- Developed `Operations.cs` for generating wine type data in JSON and C# enum formats.
- Added `WineType.cs` and `WineTypeEntry.cs` for defining wine types.
- Included `CreatePopulate.sql` for database and table creation with initial data.
- Introduced `DirectoryHelper.cs` and `SqlHelper.cs` for project folder and SQL command handling.
- Added new PNG and ICO files for UI branding.
- Created `GlobalUsings.cs` to manage global using directives.
- Established `WineToJsonApp` namespace for modular organization.

by Karen

+++ EnumHasConversionSample/Program.cs
@@ -1,4 +1,5 @@
-﻿using EnumHasConversionSample.Classes;
+﻿using ConsoleConfigurationLibrary.Classes;
+using EnumHasConversionSample.Classes;
 using EnumHasConversionSample.Data;
 
 namespace EnumHasConversionSample;
@@ -9,8 +10,17 @@ internal partial class Program
     {
         using var context = new WineContext();
 
-        context.Database.EnsureDeleted();
-        context.Database.EnsureCreated();
+        if ( EntitySettings.Instance.CreateNew == false)
+        {
+            
+        }
+
+        if (EntitySettings.Instance.CreateNew)
+        {
+            context.Database.EnsureDeleted();
+            context.Database.EnsureCreated();
+        }
+
         WineOperations.Run();
 
         Console.WriteLine("Finished");
